+++ spec/rubocop/cop/layout/heredoc_argument_closing_parenthesis_spec.rb
@@ -64,7 +64,7 @@ RSpec.describe RuboCop::Cop::Layout::HeredocArgumentClosingParenthesis, :config 
       RUBY
     end
 
-    it 'accepts when there is an argument between a heredoc argument and the closing paretheses' do
+    it 'accepts when there is an argument between a heredoc argument and the closing parentheses' do
       expect_no_offenses(<<~RUBY)
         foo(<<~TEXT,
             Lots of